+++ src/Intersection.h
@@ -23,10 +23,10 @@ public:
 
 class Intersection {
 public:
-    float t;
+    double t;
     std::unique_ptr<Geometry> geometry;
 
-    Intersection(float t, std::unique_ptr<Geometry>&& geom);
+    Intersection(double t, std::unique_ptr<Geometry>&& geom);
     Intersection(const Intersection& copy);
     auto operator=(const Intersection& rhs) -> Intersection&;
     auto operator==(const Intersection& rhs) const -> bool;
