+++ src/Material.cpp
@@ -21,13 +21,13 @@ auto Lighting(
     Color black{0,0,0};
     Color diffuse{black};
     Color specular{black};
-    float light_dot_normal{lightv.Dot(normalv)};
+    double light_dot_normal{lightv.Dot(normalv)};
     if (light_dot_normal >= 0) {
         diffuse = effective_color * mat.diffuse * light_dot_normal;
         Vector reflectv{Vector{-lightv}.Reflect(normalv)};
-        float reflect_dot_eye{reflectv.Dot(eyev)};
+        double reflect_dot_eye{reflectv.Dot(eyev)};
         if (reflect_dot_eye > 0) {
-            float factor{std::pow(reflect_dot_eye, mat.shininess)};
+            double factor{std::pow(reflect_dot_eye, mat.shininess)};
             specular = light.intensity * mat.specular * factor;
         }
     }
