+++ src/Transformations.cpp
@@ -7,24 +7,24 @@
 using std::cos;
 using std::sin;
 
-auto Transformations::Translation(const float& tx, const float& ty, const float& tz) -> Matrix4f {
-    Matrix4f buffer{Identity<float, 4>()};
+auto Transformations::Translation(const double& tx, const double& ty, const double& tz) -> Matrix4d {
+    Matrix4d buffer{Identity<double, 4>()};
     buffer(0,3) = tx;
     buffer(1,3) = ty;
     buffer(2,3) = tz;
     return buffer;
 }
 
-auto Transformations::Scaling(const float &sx, const float &sy, const float &sz) -> Matrix4f {
-    Matrix4f buffer{Identity<float, 4>()};
+auto Transformations::Scaling(const double &sx, const double &sy, const double &sz) -> Matrix4d {
+    Matrix4d buffer{Identity<double, 4>()};
     buffer(0,0) = sx;
     buffer(1,1) = sy;
     buffer(2,2) = sz;
     return buffer;
 }
 
-auto Transformations::RotationX(const float& rad) -> Matrix4f {
-    Matrix4f buffer{Identity<float, 4>()};
+auto Transformations::RotationX(const double& rad) -> Matrix4d {
+    Matrix4d buffer{Identity<double, 4>()};
     buffer(1,1) = std::cos(rad);
     buffer(1,2) = -std::sin(rad);
     buffer(2,1) = std::sin(rad);
@@ -32,8 +32,8 @@ auto Transformations::RotationX(const float& rad) -> Matrix4f {
 
     return buffer;
 }
-auto Transformations::RotationY(const float& rad) -> Matrix4f {
-    Matrix4f buffer{Identity<float, 4>()};
+auto Transformations::RotationY(const double& rad) -> Matrix4d {
+    Matrix4d buffer{Identity<double, 4>()};
     buffer(0,0) = std::cos(rad);
     buffer(0,2) = std::sin(rad);
     buffer(2,0) = -std::sin(rad);
@@ -41,8 +41,8 @@ auto Transformations::RotationY(const float& rad) -> Matrix4f {
 
     return buffer;
 }
-auto Transformations::RotationZ(const float& rad) -> Matrix4f {
-    Matrix4f buffer{Identity<float, 4>()};
+auto Transformations::RotationZ(const double& rad) -> Matrix4d {
+    Matrix4d buffer{Identity<double, 4>()};
     buffer(0,0) = std::cos(rad);
     buffer(0,1) = -std::sin(rad);
     buffer(1,0) = std::sin(rad);
@@ -51,10 +51,10 @@ auto Transformations::RotationZ(const float& rad) -> Matrix4f {
     return buffer;
 }
 
-auto Transformations::Shearing(const float& xy, const float& xz,
-                               const float& yx, const float& yz,
-                               const float& zx, const float& zy) -> Matrix4f {
-    Matrix4f buffer{Identity<float, 4>()};
+auto Transformations::Shearing(const double& xy, const double& xz,
+                               const double& yx, const double& yz,
+                               const double& zx, const double& zy) -> Matrix4d {
+    Matrix4d buffer{Identity<double, 4>()};
     buffer(0,1) = xy;
     buffer(0,2) = xz;
     buffer(1,0) = yx;
@@ -64,11 +64,11 @@ auto Transformations::Shearing(const float& xy, const float& xz,
     return buffer;
 }
 
-auto ViewTransform(const Point& from, const Point& to, const Vector& up) -> Matrix4f {
+auto ViewTransform(const Point& from, const Point& to, const Vector& up) -> Matrix4d {
     Vector forward{(to - from).Normalize()};
     Vector left{forward.Cross(up.Normalize())};
     Vector true_up{left.Cross(forward)};
-    Matrix4f orientation{{
+    Matrix4d orientation{{
         left.x,left.y,left.z,0.f,
         true_up.x,true_up.y,true_up.z,0.f,
         -forward.x,-forward.y,-forward.z,0.f,
